Extract button components out of Signer
