+++ src/components/onboarding/OnBoardingModal/steps/Signer/index.tsx
@@ -1,0 +1,3 @@
+import Signer from './Signer'
+
+export default Signer
